requirements.txt
absl-py==0.13.0
adabelief-pytorch==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
einops==0.3.0
fire==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
idna==3.2
kiwisolver==1.3.1
kornia==0.5.7
Markdown==3.3.4
matplotlib
oauthlib==3.1.1
pandas
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-gan-metrics==0.3.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
